requirements.txt
alabaster==0.7.12
Babel==2.9.0
biopython==1.78
blinker==1.4
bokeh==2.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
commonmark==0.9.1
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
dnspython==2.1.0
docutils==0.16
et-xmlfile==1.0.1
ete3==3.1.2
Faker==6.6.1
Flask==1.1.2
Flask-Cors==3.0.10
greenlet==1.0.0
idna==2.10
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
networkit==8.1
networkx==2.5
numpy==1.20.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
Pillow==8.1.2
progressbar2==3.53.1
psutil==5.8.0
Pygments==2.8.1
pymongo==3.11.3
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
Pyvolve==1.0.3
PyYAML==5.4.1
recommonmark==0.7.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
sentry-sdk==1.0.0
sip==6.0.3
six==1.15.0
snowballstemmer==2.1.0
Sphinx==3.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.1
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
xlrd==2.0.1
